Preamble Update
Added ContinueButton Script. This is more-so a change scene button. Put the path of the scene in the editor and the button will then take you there. It was a small thing that I think can help us a lot in the future. It sticks to the highlighting rules we already have in place.
Create a theme for the font. This will help with blanket changes to ui scenes.
Added a preamble scene for lore before playing the game
Redid main menu art
Increased size of story event display
Added 4 more story events to test and use